reorganize the cli code
